Refactor transaction price to an Amount method
This commit is the first step to migrate from (Quantity, Price)
to (Quantity, Amount) for transactions.

It introduces Amount() and Price() methods on the Buy and Sell
structs. The direct access to the `Price` field has been
replaced with the new methods. This will allow to safely change
the underlying data structure in the next phase of the migration.

Fixes #38