Refactor documentation tests to support extensions
This commit refactors the documentation testing to support subcommand extensions as part of the work on issue #35.

The existing test framework for `README.md` and `UserManual.md` is replaced by a more generic one that can run commands from any markdown file. This is a preliminary step to be able to have many more documentation files.

The main changes are:
- Use `goldmark` to parse markdown files and extract runnable code blocks.
- A new `documentation_test.go` to test all markdown files.
- The old `readme_test.go` and `user_manual_test.go` are removed.
- `test_helper.go` is updated with the new logic.

This new testing framework will allow us to add testable documentation easily.

Closes #35