Imply update option for today's reports
Implement the implied '-u' option when the '-d' option is set to today for
daily, gains, holding, and summary reports. This ensures that reports for
the current day are always updated with the latest intraday prices, as
today's report is meaningless without updated data.

This change involved implementing the IsToday() method in the date package
and verifying that the existing logic in the command files correctly
utilizes this method.

Closes #59